requirements.txt
paho-mqtt~=2.1.0
discord-webhook~=1.3.1